+++ src/Schema/Factories/ArgumentFactory.php
@@ -3,39 +3,57 @@
 namespace Nuwave\Lighthouse\Schema\Factories;
 
 use Nuwave\Lighthouse\Schema\AST\ASTHelper;
-use Nuwave\Lighthouse\Schema\Values\ArgumentValue;
+use GraphQL\Language\AST\InputValueDefinitionNode;
+use Nuwave\Lighthouse\Schema\Conversion\DefinitionNodeConverter;
 
 class ArgumentFactory
 {
     /**
-     * Convert argument definition to type.
+     * Convert input value definitions to a executable types.
      *
-     * @param  \Nuwave\Lighthouse\Schema\Values\ArgumentValue  $argumentValue
-     * @return array
+     * @param  \GraphQL\Language\AST\InputValueDefinitionNode[]|\GraphQL\Language\AST\NodeList  $definitions
+     * @return mixed[]
      */
-    public function handle(ArgumentValue $argumentValue): array
+    public function toTypeMap($definitionNodes): array
     {
-        $definition = $argumentValue->getAstNode();
+        $arguments = [];
 
-        $argumentType = $argumentValue->getType();
+        /* @var InputValueDefinitionNode $inputValueDefinitionNode */
+        foreach ($definitionNodes as $inputDefinition) {
+            $arguments[$inputDefinition->name->value] = $this->convert($inputDefinition);
+        }
 
-        $fieldArgument = [
-            'name' => $argumentValue->getName(),
-            'description' => data_get($definition->description, 'value'),
-            'type' => $argumentType,
-            'astNode' => $definition,
+        return $arguments;
+    }
+
+    /**
+     * Convert an argument definition to an executable type.
+     *
+     * The returned array will be used to construct one of:
+     * @see \GraphQL\Type\Definition\FieldArgument
+     * @see \GraphQL\Type\Definition\InputObjectField
+     *
+     * @param  \GraphQL\Language\AST\InputValueDefinitionNode  $definitionNode
+     * @return mixed[]
+     */
+    public function convert(InputValueDefinitionNode $definitionNode): array
+    {
+        $definitionNodeConverter = app(DefinitionNodeConverter::class);
+        $type = $definitionNodeConverter->toType($definitionNode->type);
+
+        $config = [
+            'name' => $definitionNode->name->value,
+            'description' => data_get($definitionNode->description, 'value'),
+            'type' => $type,
+            'astNode' => $definitionNode,
         ];
 
-        if ($defaultValue = $definition->defaultValue) {
-            $fieldArgument += [
-                'defaultValue' => ASTHelper::defaultValueForArgument($defaultValue, $argumentType),
+        if ($defaultValue = $definitionNode->defaultValue) {
+            $config += [
+                'defaultValue' => ASTHelper::defaultValueForArgument($defaultValue, $type),
             ];
         }
 
-        // Add any dynamically declared public properties of the FieldArgument
-        $fieldArgument += get_object_vars($argumentValue);
-
-        // Used to construct a FieldArgument class
-        return $fieldArgument;
+        return $config;
     }
 }